+++ common/components/InputComponent.hpp
@@ -7,9 +7,9 @@ namespace kengine {
     struct InputComponent : SerializableComponent<InputComponent> {
 		const std::string type = pmeta_nameof(InputComponent);
         std::function<void(int keycode, bool pressed)> onKey = nullptr;
-        std::function<void(int x, int y)> onMouseMove = nullptr;
-        std::function<void(int button, int x, int y, bool pressed)> onMouseButton = nullptr;
-        std::function<void(int delta, int x, int y)> onMouseWheel = nullptr;
+        std::function<void(float x, float y)> onMouseMove = nullptr;
+        std::function<void(int button, float x, float y, bool pressed)> onMouseButton = nullptr;
+        std::function<void(float delta, float x, float y)> onMouseWheel = nullptr;
 
         pmeta_get_class_name(InputComponent);
         pmeta_get_attributes(
